+++ components/camel-consul/src/test/java/org/apache/camel/component/consul/ConsulCoordinatesIT.java
@@ -26,7 +26,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-public class ConsulCoordinatesTest extends ConsulTestSupport {
+public class ConsulCoordinatesIT extends ConsulTestSupport {
 
     @Test
     public void testDatacenters() {
