+++ components/camel-consul/src/test/java/org/apache/camel/component/consul/ConsulRegistryIT.java
@@ -38,7 +38,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 /**
  * Unit test for Camel Registry implementation for Consul
  */
-public class ConsulRegistryTest implements Serializable {
+public class ConsulRegistryIT implements Serializable {
     @RegisterExtension
     public static ConsulService consulService = ConsulServiceFactory.createService();
