+++ components/camel-consul/src/test/java/org/apache/camel/component/consul/cloud/ConsulServiceRegistrationWithServiceComponentIT.java
@@ -16,19 +16,37 @@
  */
 package org.apache.camel.component.consul.cloud;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import org.apache.camel.BindToRegistry;
 import org.apache.camel.RoutesBuilder;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.impl.cloud.ServiceRegistrationRoutePolicy;
+import org.apache.camel.component.service.ServiceComponent;
 
-public class ConsulServiceRegistrationWithRoutePolicyTest extends ConsulServiceRegistrationTestBase {
+public class ConsulServiceRegistrationWithServiceComponentIT extends ConsulServiceRegistrationTestBase {
+
+    @BindToRegistry("service")
+    ServiceComponent comp = new ServiceComponent();
+
+    @Override
+    protected Map<String, String> getMetadata() {
+        return new HashMap<String, String>() {
+            {
+                put("service.type", "consul");
+                put("service.zone", "US");
+            }
+        };
+    }
+
     @Override
     protected RoutesBuilder createRouteBuilder() {
         return new RouteBuilder() {
             @Override
             public void configure() {
-                fromF("jetty:http://0.0.0.0:%d/service/endpoint", SERVICE_PORT).routeId(SERVICE_ID).routeGroup(SERVICE_NAME)
-                        .routePolicy(new ServiceRegistrationRoutePolicy())
-                        .noAutoStartup().to("log:service-registry?level=INFO");
+                fromF("service:%s:jetty:http://0.0.0.0:%d/service/endpoint?service.type=consul&service.zone=US", SERVICE_NAME,
+                        SERVICE_PORT).routeId(SERVICE_ID)
+                                .routeGroup(SERVICE_NAME).noAutoStartup().to("log:service-registry?level=INFO");
             }
         };
     }
