+++ components/camel-consul/src/test/java/org/apache/camel/component/consul/cloud/SpringConsulExpressionServiceCallRouteIT.java
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.context.support.AbstractApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-public class SpringConsulExpressionServiceCallRouteTest extends SpringConsulServiceCallRouteTest {
+public class SpringConsulExpressionServiceCallRouteIT extends SpringConsulServiceCallRouteTest {
     @Override
     protected AbstractApplicationContext createApplicationContext() {
         return new ClassPathXmlApplicationContext(
